+++ cli/package.json
@@ -36,15 +36,15 @@
   "dependencies": {
     "@coral-xyz/anchor": "0.29.0",
     "@lightprotocol/compressed-token": "workspace:*",
-    "@lightprotocol/stateless.js": "workspace:*",
     "@lightprotocol/hasher.rs": "workspace:*",
+    "@lightprotocol/stateless.js": "workspace:*",
     "@oclif/core": "^3.26.2",
     "@oclif/plugin-autocomplete": "^3.0.13",
     "@oclif/plugin-help": "^6.0.20",
     "@oclif/plugin-not-found": "^3.1.2",
     "@oclif/plugin-plugins": "^5.0.7",
     "@solana-developers/helpers": "^1.5.1",
-    "@solana/web3.js": "^1.95.3",
+    "@solana/web3.js": "1.95.3",
     "axios": "^1.6.8",
     "case-anything": "^2.1.13",
     "cli-progress": "^3.12.0",
@@ -73,7 +73,7 @@
     "@typescript-eslint/eslint-plugin": "^7.6.0",
     "@typescript-eslint/parser": "^7.6.0",
     "chai": "^4.4.1",
-    "eslint": "8.57.0",
+    "eslint": "9.16.0",
     "eslint-config-oclif": "5.1.1",
     "eslint-config-oclif-typescript": "3.1.4",
     "eslint-config-prettier": "9.1.0",
@@ -108,19 +108,19 @@
     "lint:fix": "eslint . --fix",
     "postpack": "shx rm -f oclif.manifest.json",
     "prepack": "pnpm build && oclif manifest && oclif readme",
-    "test-utils": "mocha ./test/utils/index.test.ts -t 10000000 --exit",
-    "test-config": "mocha ./test/commands/config/index.test.ts -t 10000000 --exit",
-    "test-create-mint": "mocha ./test/commands/create-mint/index.test.ts -t 10000000 --exit",
-    "test-merge-token-accounts": "mocha ./test/commands/merge-token-accounts/index.test.ts -t 10000000 --exit",
-    "test-create-token-pool": "mocha ./test/commands/create-token-pool/index.test.ts -t 10000000 --exit",
-    "test-approve-and-mint-to": "mocha ./test/commands/approve-and-mint-to/index.test.ts -t 10000000 --exit",
-    "test-mint-to": "mocha ./test/commands/mint-to/index.test.ts -t 10000000 --exit",
-    "test-transfer": "mocha ./test/commands/transfer/index.test.ts -t 10000000 --exit",
-    "test-balance": "mocha ./test/commands/balance/index.test.ts -t 10000000 --exit",
-    "test-compress-sol": "mocha ./test/commands/compress-sol/index.test.ts -t 10000000 --exit",
-    "test-decompress-sol": "mocha ./test/commands/decompress-sol/index.test.ts -t 10000000 --exit",
-    "test-compress-spl": "mocha ./test/commands/compress-spl/index.test.ts -t 10000000 --exit",
-    "test-decompress-spl": "mocha ./test/commands/decompress-spl/index.test.ts -t 10000000 --exit",
+    "test-utils": "mocha test/utils/index.test.ts -t 10000000 --exit",
+    "test-config": "mocha test/commands/config/index.test.ts -t 10000000 --exit",
+    "test-create-mint": "mocha test/commands/create-mint/index.test.ts -t 10000000 --exit",
+    "test-merge-token-accounts": "mocha test/commands/merge-token-accounts/index.test.ts -t 10000000 --exit",
+    "test-create-token-pool": "mocha test/commands/create-token-pool/index.test.ts -t 10000000 --exit",
+    "test-approve-and-mint-to": "mocha test/commands/approve-and-mint-to/index.test.ts -t 10000000 --exit",
+    "test-mint-to": "mocha test/commands/mint-to/index.test.ts -t 10000000 --exit",
+    "test-transfer": "mocha test/commands/transfer/index.test.ts -t 10000000 --exit",
+    "test-balance": "mocha test/commands/balance/index.test.ts -t 10000000 --exit",
+    "test-compress-sol": "mocha test/commands/compress-sol/index.test.ts -t 10000000 --exit",
+    "test-decompress-sol": "mocha test/commands/decompress-sol/index.test.ts -t 10000000 --exit",
+    "test-compress-spl": "mocha test/commands/compress-spl/index.test.ts -t 10000000 --exit",
+    "test-decompress-spl": "mocha test/commands/decompress-spl/index.test.ts -t 10000000 --exit",
     "kill": "killall solana-test-validator || true && killall solana-test-val || true && sleep 1",
     "test-cli": "pnpm test-config && pnpm kill",
     "test": "pnpm kill && pnpm test-cli && pnpm test-utils && pnpm test-create-mint && pnpm test-mint-to && pnpm test-transfer && pnpm test-merge-token-accounts && pnpm test-create-token-pool && pnpm test-compress-spl && pnpm test-decompress-spl && pnpm test-balance && pnpm test-compress-sol && pnpm test-decompress-sol && pnpm test-approve-and-mint-to",
