+++ js/stateless.js/src/utils/index.ts
@@ -7,3 +7,4 @@ export * from './send-and-confirm';
 export * from './sleep';
 export * from './test-utils';
 export * from './validation';
+export * from './calculate-compute-unit-price';
